created thumbnails
made a thumbnailing thing with some arbitrary kind of ugly thumbnail generator.
fixed some issues with the calender renderer not handling end positions correctly, like if an event ended on exactly 12:00 am, it would show a little dink on the next day.
changed the menu a bit for more organization.
added test data generator as an optional thing.
closes #3